Update: Migrate blurRadius to FloatPref

+++ Omega/src/com/saggitt/omega/blur/BlurWallpaperFilter.kt
@@ -28,7 +28,8 @@ class BlurWallpaperFilter(private val context: Context) : WallpaperFilter {
     private var blurRadius = 25
 
     override fun applyPrefs(prefs: OmegaPreferences) {
-        blurRadius = prefs.themeBlurRadius.onGetValue() / BlurWallpaperProvider.DOWNSAMPLE_FACTOR
+        blurRadius = (prefs.themeBlurRadius.onGetValue() / BlurWallpaperProvider.DOWNSAMPLE_FACTOR)
+            .toInt()
         blurRadius = blurRadius.coerceAtLeast(1).coerceAtMost(25)
     }
